exercise-app-backend/exercise_project/requirements.txt
asgiref==3.8.1
authentication==1.1.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
django-dotenv==1.4.2
djangorestframework==3.15.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
load-dotenv==0.1.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
passlib==1.7.4
pycryptodome==3.20.0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.0.7
sqlparse==0.5.0
tzdata==2024.1
Werkzeug==3.0.3
WTForms==3.1.2
